+++ authors/2024-10.json
@@ -1,0 +1,1 @@
+["wapas", "gggkik", "bnb2011", "swoon"]
